OmniPanel: Fix material bake Python errors.
The material baking functionality no longer works and
fails with a Python error, because the addon code
is attempting to set the render settings tile_x and
tile_y members, which were recently removed.  The addon
was setting these as an optimization, using smaller
tile sizes when rendering on the CPU. According to
Patrick Mours, Cycles X now defaults to progressive
rendering and does not use tiles anymore, which is
why the settings were removed. There is still a
fallback that uses tiles when rendering very large
resolution images, but configuring that uses
different settings in Python. However, setting tiles
now applies to very large images (e.g. > 8000x8000)
which is not our use case.  I've removed the calls
to set the tiles, and the baking appears to work.